Implement weighted queue pools
Priority weighting can be enabled via a config option. This will switch from original behaviour in which the plugin will only ever send the first player in the highest priority queue to the server to new logic.

The new logic will select a queue associated to the slot that has an available space randomly, where each queue is weighted according to its priority. This allows for setups in which lower priority players are occasionally sent to the destination server, in a frequency proportional to the ratio between their priority values.

Needs to be tested and documented, as the meaning of the priority value in the config is essentially reversed.